vk: Refactor swapchain into separate platform headers
