+++ rpcs3/Emu/CMakeLists.txt
@@ -600,6 +600,7 @@ if(TARGET 3rdparty_vulkan)
         RSX/VK/vkutils/instance.cpp
         RSX/VK/vkutils/scratch.cpp
         RSX/VK/vkutils/sync.cpp
+        RSX/VK/vkutils/swapchain.cpp
         RSX/VK/vkutils/memory.cpp
         RSX/VK/vkutils/device.cpp
         RSX/VK/vkutils/sampler.cpp
