+++ rpcs3/Emu/RSX/VK/vkutils/swapchain_win32.hpp
@@ -1,0 +1,98 @@
+#pragma once
+
+#include "swapchain_core.h"
+
+namespace vk
+{
+#if defined(_WIN32)
+	class swapchain_WIN32 : public native_swapchain_base
+	{
+		HDC hDstDC = NULL;
+		HDC hSrcDC = NULL;
+		HBITMAP hDIB = NULL;
+		LPVOID hPtr = NULL;
+
+	public:
+		swapchain_WIN32(physical_device& gpu, u32 present_queue, u32 graphics_queue, u32 transfer_queue, VkFormat format = VK_FORMAT_B8G8R8A8_UNORM)
+			: native_swapchain_base(gpu, present_queue, graphics_queue, transfer_queue, format)
+		{}
+
+		~swapchain_WIN32() {}
+
+		bool init() override
+		{
+			if (hDIB || hSrcDC)
+				destroy(false);
+
+			RECT rect;
+			GetClientRect(window_handle, &rect);
+			m_width = rect.right - rect.left;
+			m_height = rect.bottom - rect.top;
+
+			if (m_width == 0 || m_height == 0)
+			{
+				rsx_log.error("Invalid window dimensions %d x %d", m_width, m_height);
+				return false;
+			}
+
+			BITMAPINFO bitmap = {};
+			bitmap.bmiHeader.biSize = sizeof(BITMAPINFOHEADER);
+			bitmap.bmiHeader.biWidth = m_width;
+			bitmap.bmiHeader.biHeight = m_height * -1;
+			bitmap.bmiHeader.biPlanes = 1;
+			bitmap.bmiHeader.biBitCount = 32;
+			bitmap.bmiHeader.biCompression = BI_RGB;
+
+			hSrcDC = CreateCompatibleDC(hDstDC);
+			hDIB = CreateDIBSection(hSrcDC, &bitmap, DIB_RGB_COLORS, &hPtr, NULL, 0);
+			SelectObject(hSrcDC, hDIB);
+			init_swapchain_images(dev, 3);
+			return true;
+		}
+
+		void create(display_handle_t& handle) override
+		{
+			window_handle = handle;
+			hDstDC = GetDC(handle);
+		}
+
+		void destroy(bool full = true) override
+		{
+			DeleteObject(hDIB);
+			DeleteDC(hSrcDC);
+			hDIB = NULL;
+			hSrcDC = NULL;
+
+			swapchain_images.clear();
+
+			if (full)
+			{
+				ReleaseDC(window_handle, hDstDC);
+				hDstDC = NULL;
+
+				dev.destroy();
+			}
+		}
+
+		VkResult present(VkSemaphore /*semaphore*/, u32 image) override
+		{
+			auto& src = swapchain_images[image];
+			GdiFlush();
+
+			if (hSrcDC)
+			{
+				memcpy(hPtr, src.second->get_pixels(), src.second->get_required_memory_size());
+				BitBlt(hDstDC, 0, 0, m_width, m_height, hSrcDC, 0, 0, SRCCOPY);
+				src.second->free_pixels();
+			}
+
+			src.first = false;
+			return VK_SUCCESS;
+		}
+	};
+
+	using swapchain_NATIVE = swapchain_WIN32;
+
+
+#endif
+}
